feat: update product page with new api
